Drag and drop but actually doing it!
This is an upgrade of the previous committed program. Now you can only drag the circle if you are selecting it directly. Also, it'll be dragged from the point where it's being picked. If you lose the circle, you can press "R" (or "r") to put it in the center of the window.